requirements.txt
pandas
kafka-python
pytest